backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.9",
    "django-cors-headers>=4.9.0",
    "django-filter>=2.2.0",
    "django-jalali>=7.4.0",
    "django-unfold>=0.74.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "inflection>=0.5.1",
    "markdown>=3.10",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "uritemplate>=4.2.0",
]
